+++ bootique-jersey-client/src/main/java/io/bootique/jersey/client/HttpClientFactoryFactory.java
@@ -25,6 +25,7 @@ import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
 import io.bootique.jersey.client.auth.AuthenticatorFactory;
 import io.bootique.jersey.client.log.JULSlf4jLogger;
+import io.bootique.jersey.client.log.RequestLoggingFilter;
 import org.glassfish.hk2.api.InjectionResolver;
 import org.glassfish.jersey.client.ClientConfig;
 import org.glassfish.jersey.client.ClientProperties;
@@ -197,13 +198,11 @@ public class HttpClientFactoryFactory {
 
     protected void configRequestLogging(ClientConfig config) {
 
-        Logger logger = LoggerFactory.getLogger(HttpClientFactory.class);
+        Logger logger = LoggerFactory.getLogger(RequestLoggingFilter.class);
         if (logger.isDebugEnabled()) {
 
-            JULSlf4jLogger julWrapper = new JULSlf4jLogger(HttpClientFactoryFactory.class.getName(), logger);
-
-            LoggingFeature loggingFeature = new LoggingFeature(julWrapper);
-            config.register(loggingFeature);
+            RequestLoggingFilter logFilter = new RequestLoggingFilter();
+            config.register(logFilter);
         }
     }