+++ bootique-jersey-client/src/test/java/io/bootique/jersey/client/HttpClientFactoryFactory_LoggingIT.java
@@ -109,10 +109,8 @@ public class HttpClientFactoryFactory_LoggingIT {
 
         File log = new File(logsDir, "debug.log");
         List<String> lines = Files.readAllLines(log.toPath());
-        assertEquals(lines.stream().collect(joining("\n")), 11, lines.size());
-        assertTrue(lines.get(0).contains("Sending client request on thread main"));
-        assertTrue(lines.get(1).contains("GET http://127.0.0.1:8080/get"));
-        assertTrue(lines.get(3).contains("Client response received on thread main"));
+        assertEquals(lines.stream().collect(joining("\n")), 5, lines.size());
+        assertTrue(lines.get(3).contains("GET http://127.0.0.1:8080/get"));
     }
 
     @Ignore